server/todo/urls.py
from . import views
from django.urls import include, path
from rest_framework import routers

from .views import TodoViewSet

router = routers.DefaultRouter()
router.register(r'todo', TodoViewSet)

urlpatterns = [
    path('', include(router.urls)),
]
