server/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Markdown==3.2.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
